Changing small energy fill generation rules.
This reworks how the Witness junk pool is generated so that we can have precise control over the ratio of energy fill items in the game, regardless of the number of other items in the pool. Previously, junk items were generated only in the unfilled slots, which meant that the number of each kind of fill would vary based on both the total size of the pool and the number of non-junk checks in the game. This led to dramatic swings in the energy economy between different doors modes, since a large non-junk item pool would result in few large fills while a small non-junk pool would result in many large fills.

This change moves to a solution that is dependent primarily on total pool size and minimizes the effect of non-junk items. Since every non-junk check awards a small fill, we now treat those checks as small fills when determining the composition of the junk item pool. In other words, if the filler pool is configured such that 25% of filler is to be a large fill, we will attempt to add large fills such that 25% of the total item pool consist of large fills, then add small fills until the item pool is full. This both allows us finer control over the energy economy as well as keeps the number of small fills--which are unsatisfying to find--to a minimum.